Refactor RecallArea into Subclasses and Concrete Impls
