+++ NonMaps/src/com/nonlinearlabs/NonMaps/client/world/overlay/belt/parameters/recall/ParameterRecallArea.java
@@ -1,0 +1,22 @@
+package com.nonlinearlabs.NonMaps.client.world.overlay.belt.parameters.recall;
+
+import com.nonlinearlabs.NonMaps.client.dataModel.editBuffer.BasicParameterModel;
+import com.nonlinearlabs.NonMaps.client.dataModel.editBuffer.EditBufferModel;
+import com.nonlinearlabs.NonMaps.client.world.overlay.belt.parameters.BeltParameterLayout;
+
+public class ParameterRecallArea extends RecallArea {
+
+	public ParameterRecallArea(BeltParameterLayout parent) {
+		super(parent);
+		addChild(button = new ParameterRecallButton(this));
+		addChild(value = new ParameterRecallValue(this));
+	}
+	
+	@Override
+	public boolean isChanged() {
+		EditBufferModel eb = EditBufferModel.get();
+		BasicParameterModel bpm = eb.findParameter(eb.selectedParameter.getValue());
+		return bpm.value.value.getValue() != bpm.originalValue.getValue();
+	}
+
+}
